requirements.txt
Flask==2.1.3
Flask-Cors==3.0.10
Flask-SQLAlchemy
marshmallow==3.16.0
PyJWT==2.4.0
flask-restx